check cache_dir existance before removal

+++ app/models/concerns/folio/clears_page_cache.rb
@@ -12,7 +12,7 @@ module Folio
       return unless self.class.clears_page_cache_on_save?
       return unless Rails.application.config.action_controller.perform_caching
       cache_dir = Rails.application.config.action_controller.page_cache_directory
-      return if cache_dir.blank?
+      return if cache_dir.blank? || !File.exist?(cache_dir)
       Dir.mktmpdir { |tmp_dir| FileUtils.mv cache_dir, tmp_dir }
     end
